Added progress outputs and made script non-locking
I added a check that looks for previously exported depth maps. If depth maps were already exported in the selected output folder, they are not re-exported. I've also added print statements that display progress and what was done for each image (skipped because it's not aligned, skipped because it's already exported, or depth maps exported). The script was also made non-locking, so that these statements appear and Metashape doesn't appear to freeze during processing which can take quite a bit of time.